fix: handle single arch images
go-containerregistry's partial.Descriptor only copies Platform onto the index entry
when the image was resolved from a multi-arch index. Derive platform from the image's
config file if present so single arch images keep working.

Part of #395
